feat: Add TokenScanner.swift file
